api/requirements.txt
captcha~=0.3.0
celery~=5.0.2,!=5.0.6
coverage~=5.5
cryptography~=3.4.7
Django~=3.2.5
django-cors-headers~=3.7.0
djangorestframework~=3.12.2
django-celery-email~=3.0.0
django-netfields~=1.2.2
django-pgtrigger~=2.3.0
django-prometheus~=2.1.0
dnspython~=2.1.0
httpretty~=1.0.2
psycopg2~=2.8.5
prometheus-client~=0.9.0  # added to control django-prometheus' dependency version
psl-dns~=1.0
pylibmc~=1.6.1
pyyaml~=5.4.1
requests~=2.25.0
uwsgi~=2.0.0
